+++ app/models/scenarios/scenario_403.rb
@@ -6,8 +6,8 @@ module Scenarios
     NAME = "Shadows in the Jungle"
     ALLIES = ["aus"].freeze
     AXIS = ["jap"].freeze
-    STATUS = "a"
-    VERSION = "0.2"
+    STATUS = "b"
+    VERSION = "0.3"
 
     DATE = [1942, 7, 28].freeze
     LAYOUT = [15, 23, "x"].freeze
@@ -17,7 +17,7 @@ module Scenarios
         list: [
           :uk_leader_4_1,
           :uk_leader_3_1,
-          [9, :uk_territorial_s],
+          [6, :uk_colonial_s],
           :uk_vickers_mg,
         ],
       },
@@ -40,7 +40,7 @@ module Scenarios
     class << self
       def generate
         {
-          turns: 7,
+          turns: 6,
           first_deploy: 1,
           first_action: 2,
           date:,