+++ app/javascript/engine/control/movement.ts
@@ -101,8 +101,10 @@ export function movementCost(
     cost = 1
   }
   if (allAlongRoad(map, index) && (map.game?.moveState.path.length ?? 0) > 1 && !roadMove) {
-    if (terrFrom.move && terrTo.move && terrFrom.move > terrTo.move && hexFrom.road) {
-      cost = terrFrom.move
+    if (!target.isVehicle || hexFrom.roadType !== roadType.Path) {
+      if (terrFrom.move && terrTo.move && terrFrom.move > terrTo.move && hexFrom.road) {
+        cost = terrFrom.move
+      }
     }
   }
   if (hexFrom.border && hexFrom.borderEdges?.includes(dir)) {